+++ CHANGELOG.md
@@ -8,6 +8,8 @@ WYMeditor.
 
 * [#761](https://github.com/wymeditor/wymeditor/pull/761)
   Fix: Image resize handle below end of iframe
+* [#748](https://github.com/wymeditor/wymeditor/issues/748)
+  Reposition caret when removing row/column that contains selection
 
 ## 1.1.1